bugfix(event-sourcing) Decoupled from Core
