+++ module/api/src/Application/EventListener/RequestBodyListener.php
@@ -9,10 +9,10 @@ declare(strict_types=1);
 
 namespace Ergonode\Api\Application\EventListener;
 
+use Ergonode\SharedKernel\Application\Serializer\SerializerInterface;
 use Symfony\Component\HttpFoundation\ParameterBag;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpKernel\Event\GetResponseEvent;
-use Ergonode\Core\Application\Serializer\SerializerInterface;
 
 class RequestBodyListener
 {
