+++ module/core/src/Domain/User/UserInterface.php
@@ -10,12 +10,11 @@ declare(strict_types=1);
 namespace Ergonode\Core\Domain\User;
 
 use Ergonode\Core\Domain\ValueObject\LanguagePrivileges;
-use Ergonode\SharedKernel\Domain\AggregateId;
-use Symfony\Component\Security\Core\User\UserInterface as BaseUserInterface;
+use Ergonode\SharedKernel\Domain\User\UserInterface as DomainUserInterface;
+use Symfony\Component\Security\Core\User\UserInterface as SymfonyUserInterface;
 
-interface UserInterface extends BaseUserInterface
+interface UserInterface extends SymfonyUserInterface, DomainUserInterface
 {
-    public function getId(): AggregateId;
     public function isActive(): bool;
     /**
      * @return LanguagePrivileges[]
