+++ module/designer/tests/Domain/Factory/TemplateElementFactoryTest.php
@@ -12,9 +12,9 @@ namespace Ergonode\Designer\Tests\Domain\Factory;
 use Ergonode\Designer\Domain\Factory\TemplateElementFactory;
 use Ergonode\Designer\Domain\ValueObject\Position;
 use Ergonode\Designer\Domain\ValueObject\Size;
+use Ergonode\SharedKernel\Application\Serializer\SerializerInterface;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
-use Ergonode\Core\Application\Serializer\SerializerInterface;
 use Ergonode\Designer\Domain\Entity\TemplateElementInterface;
 use Ergonode\Designer\Domain\Entity\Element\AttributeTemplateElement;
 use Ergonode\Designer\Domain\Resolver\TemplateElementTypeResolver;
