+++ module/shared-kernel/src/Application/Serializer/Exception/NormalizationException.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * Copyright © Bold Brand Commerce Sp. z o.o. All rights reserved.
  * See LICENSE.txt for license details.
@@ -6,9 +7,9 @@
 
 declare(strict_types=1);
 
-namespace Ergonode\Core\Application\Exception;
+namespace Ergonode\SharedKernel\Application\Serializer\Exception;
 
-class DeserializationException extends SerializerException
+class NormalizationException extends NormalizerException
 {
     public function __construct(string $message, \Throwable $previous = null)
     {
