+++ module/shared-kernel/src/Application/Serializer/Exception/SerializationException.php
@@ -1,5 +1,4 @@
 <?php
-
 /**
  * Copyright © Bold Brand Commerce Sp. z o.o. All rights reserved.
  * See LICENSE.txt for license details.
@@ -7,9 +6,9 @@
 
 declare(strict_types=1);
 
-namespace Ergonode\Core\Application\Exception;
+namespace Ergonode\SharedKernel\Application\Serializer\Exception;
 
-class DenoralizationException extends NormalizerException
+class SerializationException extends SerializerException
 {
     public function __construct(string $message, \Throwable $previous = null)
     {
