+++ module/shared-kernel/src/Application/Serializer/Exception/SerializerException.php
@@ -1,5 +1,4 @@
 <?php
-
 /**
  * Copyright © Bold Brand Commerce Sp. z o.o. All rights reserved.
  * See LICENSE.txt for license details.
@@ -7,12 +6,12 @@
 
 declare(strict_types=1);
 
-namespace Ergonode\Core\Application\Exception;
+namespace Ergonode\SharedKernel\Application\Serializer\Exception;
 
-class NormalizationException extends NormalizerException
+class SerializerException extends \Exception
 {
     public function __construct(string $message, \Throwable $previous = null)
     {
-        parent::__construct($message, $previous);
+        parent::__construct($message, 0, $previous);
     }
 }
