+++ module/shared-kernel/src/Domain/User/UserInterface.php
@@ -1,0 +1,17 @@
+<?php
+
+/**
+ * Copyright © Bold Brand Commerce Sp. z o.o. All rights reserved.
+ * See LICENSE.txt for license details.
+ */
+
+declare(strict_types=1);
+
+namespace Ergonode\SharedKernel\Domain\User;
+
+use Ergonode\SharedKernel\Domain\Aggregate\UserId;
+
+interface UserInterface
+{
+    public function getId(): UserId;
+}
