+++ module/workflow/src/Infrastructure/Persistence/Projector/DbalStatusDescriptionChangedEventProjector.php
@@ -10,8 +10,8 @@ declare(strict_types=1);
 namespace Ergonode\Workflow\Infrastructure\Persistence\Projector;
 
 use Doctrine\DBAL\Connection;
+use Ergonode\SharedKernel\Application\Serializer\SerializerInterface;
 use Ergonode\Workflow\Domain\Event\Status\StatusDescriptionChangedEvent;
-use Ergonode\Core\Application\Serializer\SerializerInterface;
 
 class DbalStatusDescriptionChangedEventProjector
 {
