+++ module/workflow/src/Infrastructure/Persistence/Projector/DbalStatusNameChangedEventProjector.php
@@ -10,8 +10,8 @@ declare(strict_types=1);
 namespace Ergonode\Workflow\Infrastructure\Persistence\Projector;
 
 use Doctrine\DBAL\Connection;
+use Ergonode\SharedKernel\Application\Serializer\SerializerInterface;
 use Ergonode\Workflow\Domain\Event\Status\StatusNameChangedEvent;
-use Ergonode\Core\Application\Serializer\SerializerInterface;
 
 class DbalStatusNameChangedEventProjector
 {
